starting to add meta piano roll functionality
able to draw music event windows on the project grid. double-click it to
pull up the piano roll. Eventually, the piano roll that comes up will be
only as long as the event window is and show only the notes in that
window. This is probably gonna take some time to implement